Grow hashmap when it fills up

+++ include/motor/hashmap.h
@@ -8,21 +8,21 @@
 typedef struct MtArena MtArena;
 
 typedef struct MtHashMap {
-  MtArena* arena;
+  MtArena *arena;
   uint64_t *keys;
-  uint32_t *values;
-  uint32_t size;
+  uint64_t *values;
+  uintptr_t size;
 } MtHashMap;
 
-void mt_hash_init(MtHashMap *map, uint32_t size, MtArena* arena);
+void mt_hash_init(MtHashMap *map, uint32_t size, MtArena *arena);
 
 void mt_hash_clear(MtHashMap *map);
 
-uint32_t mt_hash_set(MtHashMap *map, uint64_t key, uint32_t value);
+uintptr_t mt_hash_set(MtHashMap *map, uint64_t key, uintptr_t value);
 
 void mt_hash_remove(MtHashMap *map, uint64_t key);
 
-uint32_t mt_hash_get(MtHashMap *map, uint64_t key);
+uintptr_t mt_hash_get(MtHashMap *map, uint64_t key);
 
 void mt_hash_destroy(MtHashMap *map);
